+++ Source/Documentation/Manual/news.rst
@@ -40,6 +40,7 @@ What's been improved
 - controls for electric locos
 - air brake features for both European and American brake systems 
 - accurate friction simulation for 4 types of brake shoes
+- superelevation system rewritten to be smoother and more configurable
 
 
 Other changes